# openapi.yaml
openapi: 3.1.0
info:
  title: Game Consign Public Marketplace API
  version: 0.1.0
  license:
    name: Proprietary assessment submission
    identifier: LicenseRef-Proprietary
  description: |
    Implemented contract for the hardened marketplace core. JSON field and enum
    names use English snake_case. Unless a path explicitly overrides security,
    a short-lived opaque bearer access token is required.
servers:
  - url: /
    description: Same-origin API base; local Compose listens on port 3000
...
security:
  - access_token: []
paths:
...
  /v1/checkout:
    post:
      tags: [Checkout]
      summary: Atomically consume a cart, decrement stock, and create an order
      operationId: checkout
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '201':
          description: Order created, or the stored response for an identical retry.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Idempotency-Replayed:
              schema:
                type: boolean
              description: True when this is a stored replay response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/InvalidSessionProblem'
        '409':
          description: Cart/version conflict, insufficient stock, or key reused with another request.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/RateLimitedProblem'
...
components:
  securitySchemes:
    access_token:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Short-lived opaque access token.
    refresh_cookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Rotating opaque refresh-token cookie.
    payment_signature:
      type: apiKey
      in: header
      name: X-Payment-Signature
      description: Hex-encoded HMAC-SHA256 of timestamp and exact raw body.
  headers:
    RequestId:
      schema:
        type: string
        format: uuid
      description: Server correlation identifier.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
...
  responses:
    ValidationProblem:
      description: Request validation failed.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    InvalidSessionProblem:
      description: Session is missing, expired, or revoked.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    RateLimitedProblem:
      description: Rate limit exceeded.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Money:
      type: object
      additionalProperties: false
      required: [amount_minor, currency]
      properties:
        amount_minor:
          type: integer
          format: int64
          minimum: 0
        currency:
          type: string
          const: IDR
...
    CheckoutRequest:
      type: object
      additionalProperties: false
      required: [cart_id, cart_version]
      properties:
        cart_id:
          type: string
          format: uuid
        cart_version:
          type: integer
          minimum: 1
    CheckoutResponse:
      type: object
      additionalProperties: false
      required: [order_id, order_status, payment_reference, total, created_at]
      properties:
        order_id:
          type: string
          format: uuid
        order_status:
          type: string
          const: pending_payment
        payment_reference:
          type: string
          minLength: 16
          maxLength: 100
        total:
          $ref: '#/components/schemas/Money'
        created_at:
          type: string
          format: date-time
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          maxLength: 200
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z0-9_]+$'
          examples:
            - VALIDATION_FAILED
            - INVALID_CREDENTIALS
            - OUT_OF_STOCK
            - IDEMPOTENCY_KEY_REUSED
        detail:
          type: string
          maxLength: 1000
        request_id:
          type: string
          format: uuid
        errors:
          type: array
          maxItems: 50
          items:
            type: object
            additionalProperties: false
            required: [field, code]
            properties:
              field:
                type: string
                maxLength: 200
              code:
                type: string
                maxLength: 100